Creating weapine logic
