+++ Assets/Undead Survivor/Scripts/Weapone.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: abd5ac9df2facf546bf229f59a3b260a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
