+++ Assets/Undead Survivor/Scripts/Weapone/Pistol.cs
@@ -1,0 +1,4 @@
+using UnityEngine;
+
+public class Pistol : WeaponeBaseType
+{ }
