+++ Assets/Undead Survivor/Scripts/Weapone/Pistol.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4c9bd0088fe00343a096e43981cc122
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
