+++ Assets/Undead Survivor/Scripts/Weapone/Rifle.cs
@@ -1,0 +1,4 @@
+using UnityEngine;
+
+public class Rifle : WeaponeBaseType
+{ }
