+++ Assets/Undead Survivor/Scripts/Weapone/Rifle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3398b8637906d25479f6b93d1a0d944f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
