+++ Assets/Undead Survivor/Scripts/Weapone/Shotgun.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public class Shotgun : WeaponeBaseType
+{
+    public
+}
