+++ Assets/Undead Survivor/Scripts/Weapone/Shotgun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0b4aa99419a59754784b648afe644c5e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
